FIXED sf ticket #1: "problem with writing single prec array" read/write support for single-prec floats was completely broken reported 2013-11-21 by Diego on Matlab Central
git-svn-id: https://svn.code.sf.net/p/csmatio/code/trunk@8 6a4c67c8-5740-49f5-aecb-95da1e218de0

+++ src/io/MatFileReader.cs
@@ -365,19 +365,17 @@ namespace csmatio.io
                     }
                     break;
 				case MLArray.mxSINGLE_CLASS:
-					// TODO: check: 
-					// Why ist it MLNumericArray<double> and not MLNumericArray<single> ?
 					mlArray = new MLSingle(name, dims, type, attributes);
 					//read real
 					tag = new ISMatTag(buf);
-					tag.ReadToByteBuffer(((MLNumericArray<double>)mlArray).RealByteBuffer,
+					tag.ReadToByteBuffer(((MLNumericArray<float>)mlArray).RealByteBuffer,
 						(ByteStorageSupport)mlArray);
 
 					// read complex
 					if (mlArray.IsComplex)
 					{
 						tag = new ISMatTag(buf);
-						tag.ReadToByteBuffer(((MLNumericArray<double>)mlArray).ImaginaryByteBuffer,
+						tag.ReadToByteBuffer(((MLNumericArray<float>)mlArray).ImaginaryByteBuffer,
 						(ByteStorageSupport)mlArray);
 					}
 					break;
